+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [Unreleased] - 
+
+## [a0.0.1] - 2015-12-20
 
 ### Changed
 - Removed --set command line parameter under manifest mode.
 This functionality should be handled by config mode.
+
+- Added command line parameters for all modes "pack, unpack, config, manifest"